% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interop.R
\name{as_ddi}
\alias{as_ddi}
\title{Convert XML trees to DDI objects}
\usage{
as_ddi(x, ...)
}
\arguments{
\item{x}{An \code{xml_node} object}

\item{...}{Arguments to pass to methods}
}
\value{
The DDI equivalent of the XML tree
}
\description{
Convert XML trees to DDI objects
}
