% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interop.R
\name{as_xml_string}
\alias{as_xml_string}
\title{Shortcut to text representation of DDI XML}
\usage{
as_xml_string(x, ...)
}
\arguments{
\item{x}{A ddi_node object.}

\item{...}{Arguments forwarded to \code{as_xml()}.}
}
\value{
A string containing the text representation of XML.
}
\description{
Functionally equivalent to \code{as.character(as_xml(ddi_node_obj))}.
}
