% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{create_object_id}
\alias{create_object_id}
\title{Create object IDs}
\usage{
create_object_id(
  object,
  ...,
  generator = getOption("rddi.id_generator"),
  sep = "."
)
}
\arguments{
\item{object}{A character vector containing (presumably unique) object names}

\item{...}{Arguments passed onto \code{generator}}

\item{generator}{A function that takes \code{object} and \code{...} and returns a character vector of IDs. Setting the "rddi.id_generator" option will override this functionality.}

\item{sep}{Any spaces in each element of \code{object} will be replaced with \code{sep}. Note: strings of whitespace will be replaced by a single \code{sep}, not multiple.}
}
\description{
In the DDI Codebook 2.5 specification, object IDs are not necessary,
but they are recommended to create linkages. For example, variables being
specific to a certain database file should include an ID reference to
that file.
}
\details{
To create these IDs, this function employs a generator that accepts the
\code{object} and dotted parameters to create a character vector of IDs.
By default, generated IDs adhere to the DDI 3 URN standard.
}
