% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_anlysUnit}
\alias{ddi_anlysUnit}
\title{DDI Analysis Unit specification}
\usage{
ddi_anlysUnit(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Provides information regarding whom or what the variable/nCube describes. The element may be repeated only to support multiple language expressions of the content.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/anlysUnit.html}
}

