% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_catgryGrp}
\alias{ddi_catgryGrp}
\title{DDI Category Group specification}
\usage{
ddi_catgryGrp(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
A description of response categories that might be grouped together.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/catgryGrp.html}
}

