% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_catLevel}
\alias{ddi_catLevel}
\title{DDI catLevel specification}
\usage{
ddi_catLevel(level_name, ...)
}
\arguments{
\item{level_name}{A category level name for hierarchical categorical variables (e.g. ordered categories)}

\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
Used to describe the levels of the category hierarchy. This field has no content.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/catLevel.html}
}

