% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_catValu}
\alias{ddi_catValu}
\title{DDI Category Value spcification}
\usage{
ddi_catValu(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
The value of a specific category
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/catValu.html}
}

