% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_codeBook}
\alias{ddi_codeBook}
\title{Codebook}
\usage{
ddi_codeBook(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object
}
\description{
The root node of a DDI 2.5 Codebook file. This file must contain stdyDscr.
More information on this element, especially the allowed attributes, can
be found in the references.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_dataDscr]{ddi_dataDscr()}}
\item \code{\link[=ddi_docDscr]{ddi_docDscr()}}
\item \code{\link[=ddi_fileDscr]{ddi_fileDscr()}}
\item \code{\link[=ddi_otherMat]{ddi_otherMat()}}
\item \code{\link[=ddi_stdyDscr]{ddi_stdyDscr()}}
}
}

\examples{
# All ddi_codeBook() functions must contain ddi_stdyDscr(), 
# which also has ddi_citation() as a required child element.

ddi_codeBook(ddi_stdyDscr(ddi_citation()))

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/codeBook.html}{codeBook documentation}
}
