% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_codInstr}
\alias{ddi_codInstr}
\title{DDI Coder Instructions specification}
\usage{
ddi_codInstr(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
One or more statistical measures that describe the responses to a particular variable and may include one or more standard summaries, e.g., minimum and
maximum values, median, mode, etc.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/codInstr.html}
}

