% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_cohort}
\alias{ddi_cohort}
\title{cohort and its child nodes}
\usage{
ddi_cohort(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
The element cohort is used when the nCube contains a limited number of
categories from a particular variable, as opposed to the full range of
categories. The attribute "catRef" is an IDREF to the actual category
being used. The attribute "value" indicates the actual value attached to
the category that is being used. More information on these elements,
especially their allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{cohort} is contained in \code{dmns}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_range]{ddi_range()}}
}
}

\examples{
ddi_cohort(catRef = "CV24_1", value = "1")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/cohort.html}{cohort documentation}
}
