% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccs.R
\name{ddi_dataAccs}
\alias{ddi_dataAccs}
\title{dataAccs and its children}
\usage{
ddi_dataAccs(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
This section describes data access conditions and terms of use for the data
collection. In cases where access conditions differ across individual files
or variables, multiple access conditions can be specified.  More information
on this element, especially the allowed attributes, can be found in the
references.
}
\details{
\emph{Parent node}

\code{dataAccs} is contained in \code{stdyDscr}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
\item \code{\link[=ddi_setAvail]{ddi_setAvail()}}
\item \code{\link[=ddi_useStmt]{ddi_useStmt()}}
}
}

\examples{
ddi_dataAccs()

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataAccs.html}{dataAccs documentation}
}
