% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_dataDscr}
\alias{ddi_dataDscr}
\title{dataDscr and its children}
\usage{
ddi_dataDscr(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object
}
\description{
Description of variables within the Codebook. More information on this
element, especially the allowed attributes, can be found in the references.
}
\details{
\emph{Parent node}

\code{dataDscr} is contained in \code{codeBook}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_nCube]{ddi_nCube()}}
\item \code{\link[=ddi_nCubeGrp]{ddi_nCubeGrp()}}
\item \code{\link[=ddi_notes]{ddi_notes()}}
\item \code{\link[=ddi_var]{ddi_var()}}
\item \code{\link[=ddi_varGrp]{ddi_varGrp()}}
}
}

\examples{
ddi_dataDscr()

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataDscr.html}{dataDscr documentation}
}
