% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_dataFingerprint}
\alias{ddi_dataFingerprint}
\alias{ddi_algorithmSpecification}
\alias{ddi_algorithmVersion}
\alias{ddi_digitalFingerprintValue}
\title{dataFingerprint and its child nodes}
\usage{
ddi_dataFingerprint(...)

ddi_algorithmSpecification(...)

ddi_algorithmVersion(...)

ddi_digitalFingerprintValue(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema.}
}
\description{
Allows for assigning a hash value (digital fingerprint) to the data or
data file. More information on the allowed attributes for fileTxt and
its child nodes can be found below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataFingerprint.html}{dataFingerprint documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/algorithmSpecification.html}{algorithmSpecification documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/algorithmVersion.html}{algorithmVersion documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/digitalFingerprintValue.html}{digitalFingerprintValue documentation}
}
