% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_defntn}
\alias{ddi_defntn}
\title{DDI Definition specification}
\usage{
ddi_defntn(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Rationale for why the group was constituted in this way.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/defntn.html}
}

