% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_dimensns}
\alias{ddi_dimensns}
\alias{ddi_caseQnty}
\alias{ddi_logRecL}
\alias{ddi_recNumTot}
\alias{ddi_recPrCas}
\alias{ddi_varQnty}
\alias{ddi_recDimnsn}
\title{dimensns, recDimnsn and their child nodes}
\usage{
ddi_dimensns(...)

ddi_caseQnty(...)

ddi_logRecL(...)

ddi_recNumTot(...)

ddi_recPrCas(...)

ddi_varQnty(...)

ddi_recDimnsn(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema.}
}
\description{
Dimensions of the overall digital or physical file. More information on the allowed attributes
for dimensns, recDimnsn and its child nodes can be found below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dimensns.html}{dimensns documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/recDimnsn.html}{recDimnsn documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/caseQnty.html}{caseQnty documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/logRecL.html}{logRecL documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/recNumTot.html}{recNumTot documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/recPrCas.html}{recPrCas documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/varQnty.html}{varQnty documentation}
}
