% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_distStmt}
\alias{ddi_distStmt}
\alias{ddi_depDate}
\alias{ddi_depositr}
\alias{ddi_distDate}
\alias{ddi_distrbtr}
\title{ddi_distStmt and its children}
\usage{
ddi_distStmt(...)

ddi_depDate(...)

ddi_depositr(...)

ddi_distDate(...)

ddi_distrbtr(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Distribution statement for the work at the appropriate level: marked-up document; marked-up document source; study; study
description, other material; other material for study. For more information see the references.
}
\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_contact]{ddi_contact()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/distStmt.html}{distStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/depDate.html}{depDate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/depositr.html}{depositr documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/distDate.html}{distDate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/distrbtr.html}{distrbtr documentation}
}
