% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docDscr_variables.R
\name{ddi_docSrc}
\alias{ddi_docSrc}
\title{docSrc and its child nodes}
\usage{
ddi_docSrc(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Citation for the source document. This element encodes the bibliographic
information describing the source codebook, including title information,
statement of responsibility, production and distribution information, series
and version information, text of a preferred bibliographic citation, and
notes (if any). Information for this section should be taken directly from
the source document whenever possible. If additional information is obtained
and entered in the elements within this section, the source of this
information should be noted in the source attribute of the particular
element tag. More information on this element, especially the allowed
attributes, can be found in the references.
}
\details{
\emph{Parent node}

\code{docSrc} is contained in \code{docDscr}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_biblCit]{ddi_biblCit()}}
\item \code{\link[=ddi_distStmt]{ddi_distStmt()}}
\item \code{\link[=ddi_holdings]{ddi_holdings()}}
\item \code{\link[=ddi_notes]{ddi_notes()}}
\item \code{\link[=ddi_prodStmt]{ddi_prodStmt()}}
\item \code{\link[=ddi_rspStmt]{ddi_rspStmt()}}
\item \code{\link[=ddi_serStmt]{ddi_serStmt()}}
\item \code{\link[=ddi_titlStmt]{ddi_titlStmt()}}
\item \code{\link[=ddi_verStmt]{ddi_verStmt()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/docSrc.html}{docSrc documentation}
}
