% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_drvcmd}
\alias{ddi_drvcmd}
\title{DDI Derivation Command specification}
\usage{
ddi_drvcmd(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
The actual command used to generate the derived variable. The "syntax" attribute is used to indicate the command language employed (e.g., SPSS, SAS, Fortran, etc.).
The element may be repeated to support multiple language expressions of the content.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/drvcmd.html}
}

