% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_drvdesc}
\alias{ddi_drvdesc}
\title{DDI Derivation Description specification}
\usage{
ddi_drvdesc(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
A textual description of the way in which this variable was derived. The element may be repeated to support multiple language expressions of the content.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/drvdesc.html}
}

