% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_embargo}
\alias{ddi_embargo}
\title{embargo node}
\usage{
ddi_embargo(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Provides information on variables/nCubes which are not currently available
because of policies established by the principal investigators and/or data
producers. This element may be repeated to support multiple language
expressions of the content.
}
\details{
\emph{Parent nodes}

\code{embargo} is contained in \code{nCube} and \code{var}.
}
\examples{
ddi_embargo(event = "notBefore",
            date = "2001-09-30",
            "The data associated with this variable/nCube will not become 
            available until September 30, 2001, because of embargo provisions 
            established by the data producers.")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/embargo.html}{embargo documentation}
}
