% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_exPostEvaluation}
\alias{ddi_exPostEvaluation}
\alias{ddi_evaluationProcess}
\alias{ddi_evaluator}
\alias{ddi_outcomes}
\title{exPostEvaluation and its leaf node children}
\usage{
ddi_exPostEvaluation(...)

ddi_evaluationProcess(...)

ddi_evaluator(...)

ddi_outcomes(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Post Evaluation Procedures describes evaluation procedures not address in data evaluation processes. These may include issues such as timing of the study, sequencing
issues, cost/budget issues, relevance, instituional or legal arrangments etc. of the study. More information on hte allowed attributes and child nodes can be found below
and in the references
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/exPostEvaluation.html}{exPostEvaluation documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/evaluationProcess.html}{evaluationProcess documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/evaluator.html}{evaluator documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/outcomes.html}{outcomes documentation}
}
