% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_fileStrc}
\alias{ddi_fileStrc}
\title{fileStrc and its child nodes}
\usage{
ddi_fileStrc(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Type of file structure. The file structure is fully described in the first
\code{fileTxt} within the \code{fileDscr} and then the \code{fileStrc} in subsequent
\code{fileTxt} descriptions would reference the first fileStrcRef attribute rather
than repeat the details. More information on these elements, especially
their allowed attributes, can be found in the references.
}
\details{
\emph{Parent node}

\code{fileStrc} is contained in \code{fileTxt}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
\item \code{\link[=ddi_recGrp]{ddi_recGrp()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileStrc.html}{fileStrc documentation}
}
