% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_geomap}
\alias{ddi_geomap}
\title{DDI Geographic Map specification}
\usage{
ddi_geomap(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
This element is used to point, using a "URI" attribute, to an external map that displays the geography in question. The "levelno" attribute indicates the level
of the geographic hierarchy relayed in the map. The "mapformat" attribute indicates the format of the map. This is an empty element consisting only of attributes
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/geoMap.html}
}

