% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_imputation}
\alias{ddi_imputation}
\title{DDI Imputation specification}
\usage{
ddi_imputation(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
According to the Statistical Terminology glossary maintained by the National Science Foundation, this is "the process by which one estimates missing values
for items that a survey respondent failed to provide," and if applicable in this context, it refers to the type of procedure used.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/imputation.html}
}

