% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_item}
\alias{ddi_item}
\title{DDI Value Item specification}
\usage{
ddi_item(value, ...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
The counterpart to Range; used to encode individual values. This is an empty element consisting only of its attributes. The "UNITS" attribute permits the
specification of integer/real numbers. The "VALUE" attribute is required specifies the actual value.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/item.html}
}

