% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_key}
\alias{ddi_key}
\title{DDI Range Key specification}
\usage{
ddi_key(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
This element permits a listing of the category values and labels. While this information is coded separately in the Category element, there may be some value in
having this information in proximity to the range of valid and invalid values.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/key.html}
}

