% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locMap.R
\name{ddi_locMap}
\alias{ddi_locMap}
\title{locMap and its child nodes}
\usage{
ddi_locMap(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
This element maps individual data entries to one or more physical storage
locations. It is used to describe the physical location of aggregate/tabular
data in cases where the nCube model is employed. More information on these
elements, especially their allowed attributes, can be found in the references.

\emph{Parent nodes}
}
\details{
\code{locMap} is contained in \code{fileDscr}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_dataItem]{ddi_dataItem()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/locMap.html}{locMap documentation}
}
