% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_method}
\alias{ddi_method}
\alias{ddi_dataProcessing}
\alias{ddi_stdyClas}
\title{method and its child nodes}
\usage{
ddi_method(...)

ddi_dataProcessing(...)

ddi_stdyClas(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
This section describes the methodology and processing involved in a data collection. More information on allowed attributes and
child nodes can be found below and in hte references.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_anlyInfo]{ddi_anlyInfo()}}
\item \code{\link[=ddi_codingInstructions]{ddi_codingInstructions()}}
\item \code{\link[=ddi_dataColl]{ddi_dataColl()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{ddi_notes()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/method.html}{method documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataProcessing.html}{dataProcessing documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/stdyClas.html}{stdyClas documentation}
}
