% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_mi}
\alias{ddi_mi}
\title{DDI Mathematical Identifier specification}
\usage{
ddi_mi(varRef, ...)
}
\arguments{
\item{varRef}{A required attribute consisting of a variable's IDREF}

\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
Token element containing the smallest unit in the Mathematical Row (\code{mrow()}) that carries meaning.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/mi.html}
}

