% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othrStdyMat.R
\name{ddi_othrStdyMat}
\alias{ddi_othrStdyMat}
\alias{ddi_othRefs}
\alias{ddi_relMat}
\alias{ddi_relPubl}
\alias{ddi_relStdy}
\title{othrStdyMat and its leaf node children}
\usage{
ddi_othrStdyMat(...)

ddi_othRefs(...)

ddi_relMat(...)

ddi_relPubl(...)

ddi_relStdy(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Other study description materials relating to the study description. This section describes other materials that are related to the study description that are primarily
descriptions of the content and use of the study, such as appendices, sampling information, weighting details, methodological and technical details,
publications based upon the study content, related studies or collections of studies, etc. This section may point to other materials related to the
description of the study through use of the generic citation element, which is available for each element in this section. This maps to Dublin Core
Relation element. More informaiton on othrStdyMat including its attributes and children can be found in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/othrStdyMat.html}{othrStdyMat documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/othRefs.html}{othRefs documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/relMat.html}{relMat documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/relPubl.html}{relPubl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/relStdy.html}{relStdy documentation}
}
