% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_qualityStatement}
\alias{ddi_qualityStatement}
\alias{ddi_otherQualityStatement}
\title{qualityStatement}
\usage{
ddi_qualityStatement(...)

ddi_otherQualityStatement(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
The Quality Statment consists of two parts, standardsCompliance and otherQualityStatements. In standardsCompliance list all specific standards complied with during
the execution of this study. Note the standard name and producer and how the study complied with the standard. Enter any additional quality statements in
otherQualityStatements. More information on the allowed attributes for qualityStatement and its child nodes can be found below
and in the references
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_standardsCompliance]{ddi_standardsCompliance()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/qualityStatement.html}{qualityStatment documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/otherQualityStatement.html}{otherQualityStatement documentation}
}
