% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_recGrp}
\alias{ddi_recGrp}
\title{recGrp and its child nodes}
\usage{
ddi_recGrp(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Used to describe record groupings if the file is hierarchical or relational.
More information on these elements, especially their allowed attributes,
can be found in the references.
}
\details{
\emph{Parent nodes}

\code{recGrp} is contained in \code{fileStrc}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_labl]{ddi_labl()}}
\item \code{\link[=ddi_recDimnsn]{ddi_recDimnsn()}}
}
}

\examples{
ddi_recGrp()

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/recGrp.html}{recGrp documentation}
}
