% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDevelopment.R
\name{ddi_resource}
\alias{ddi_resource}
\title{resource and its child nodes}
\usage{
ddi_resource(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Resources used in the development of the activity. More information on these
elements, especially their allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{resource} is contained in \code{developmentActivity}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_dataSrc]{ddi_dataSrc()}}
\item \code{\link[=ddi_srcChar]{ddi_srcChar()}}
\item \code{\link[=ddi_srcDocu]{ddi_srcDocu()}}
\item \code{\link[=ddi_srcOrig]{ddi_srcOrig()}}
}
}

\examples{
ddi_resource()

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/resource.html}{resource documentation}
}
