% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_respUnit}
\alias{ddi_respUnit}
\title{DDI Response Unit specification}
\usage{
ddi_respUnit(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Provides information regarding who provided the information contained within the variable/nCube, e.g., respondent, proxy, interviewer. This element may be repeated
only to support multiple language expressions of the content.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/respUnit.html}
}

