% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_rspStmt}
\alias{ddi_rspStmt}
\alias{ddi_AuthEnty}
\alias{ddi_othId}
\title{rspStmt and its child nodes}
\usage{
ddi_rspStmt(...)

ddi_AuthEnty(...)

ddi_othId(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Responsibility statement for the creation of the work at the appropriate level: marked-up document; marked-up document source; study; study description, other material; other
material for study. More information on rspStmt, its allowed attributes, and its child nodes can be found below and in the references
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/rspStmt.html}{rspStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/AuthEnty.html}{AuthEnty documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/othId.html}{othId documentation}
}
