% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_security}
\alias{ddi_security}
\title{security node}
\usage{
ddi_security(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Provides information regarding levels of access, e.g., public, subscriber,
need to know. The ISO standard for dates (YYYY-MM-DD) is recommended for use
with the date attribute. More information on this element, especially its
allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{security} is contained in \code{nCube} and \code{var}.
}
\examples{
ddi_security(date = "1998-05-10",
             "This variable has been recoded for reasons of confidentiality. 
             Users should contact the archive for information on obtaining access.")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/security.html}{security documentation}
}
