% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_serStmt}
\alias{ddi_serStmt}
\alias{ddi_serName}
\alias{ddi_serInfo}
\title{serStmt and its child nodes}
\usage{
ddi_serStmt(...)

ddi_serName(...)

ddi_serInfo(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Series statement for the work at the appropriate level: marked-up document; marked-up document source; study; study description, other material; other material
for study. More information on its child nodes and attributes can be found below.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/serStmt.html}{serStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/serName.html}{serName documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/serInfo.html}{serInfo documentation}
}
