% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccs.R
\name{ddi_setAvail}
\alias{ddi_setAvail}
\alias{ddi_accsPlac}
\alias{ddi_avlStatus}
\alias{ddi_collSize}
\alias{ddi_complete}
\alias{ddi_fileQnty}
\alias{ddi_origArch}
\title{setAvail and its children}
\usage{
ddi_setAvail(...)

ddi_accsPlac(...)

ddi_avlStatus(...)

ddi_collSize(...)

ddi_complete(...)

ddi_fileQnty(...)

ddi_origArch(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Information on availability and storage of the data set collection. For more information on setAvail and its children and attributes see the references.
}
\section{General Children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/setAvail.html}{setAvail documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/accsPlac.html}{accsPlac documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/avlStatus.html}{avlStatus documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/collSize.html}{collSize documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/complete.html}{complete documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileQnty.html}{fileQnty documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/origArch.html}{origArch documentation}
}
