% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_sources}
\alias{ddi_sources}
\title{Sources Statement}
\usage{
ddi_sources(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Description of sources used for the data collection. The element is nestable so that the sources statement might encompass a series of discrete source
statements, each of which could contain the facts about an individual source. This element maps to Dublin Core Source element.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_sourceCitation]{ddi_sourceCitation()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{ddi_dataSrc()}
\item \code{ddi_srcOrig()}
\item \code{ddi_srcChar()}
\item \code{ddi_srcDocu()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sources.html}{sources documentation}
}
