% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_srcChar}
\alias{ddi_srcChar}
\title{srcChar node}
\usage{
ddi_srcChar(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Assessment of characteristics and quality of source material. May not be
relevant to survey data. This element may be repeated to support multiple
language expressions of the content. More information on this element,
especially its allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{srcChar} is contained in the following elements: \code{sources} and \code{resource}.
}
\examples{
ddi_srcChar("Assessment of source material(s).")
 
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/srcChar.html}{srcChar documentation}
}
