% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_srcDocu}
\alias{ddi_srcDocu}
\title{srcDocu node}
\usage{
ddi_srcDocu(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Level of documentation of the original sources. May not be relevant to survey data. This element may be repeated to support multiple language expressions of the content.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/srcDocu.html}{srcDocu documentation}
}
