% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_standard}
\alias{ddi_standard}
\alias{ddi_standardName}
\title{standard}
\usage{
ddi_standard(...)

ddi_standardName(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Standard describes a standard with which the study complies. More information on the element, its attributes and its child nodes can be found
below and in the references
}
\section{General children allowed}{

*\code{\link[=ddi_producer]{ddi_producer()}}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/standard.html}{standard documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/standardName.html}{standardName documentation}
}
