% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_stdCatgry}
\alias{ddi_stdCatgry}
\title{DDI Standard Categories specification}
\usage{
ddi_stdCatgry(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Standard category codes used in the variable, like industry codes, employment codes, or social class codes
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/stdCatgry.html}
}

