% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAuthorization.R
\name{ddi_studyAuthorization}
\alias{ddi_studyAuthorization}
\alias{ddi_authorizingAgency}
\alias{ddi_authorizationStatement}
\title{studyAuthorization and its leaf node children}
\usage{
ddi_studyAuthorization(...)

ddi_authorizingAgency(...)

ddi_authorizationStatement(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Study Authorization provides structured information on the agency that authorized the study, the date of authorization, and an authorization statement.
More information on the allowed attributes its child nodes can be found in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/studyAuthorization.html}{studyAuthorization docuemntation}

\href{{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/authorizingAgency.html}}{authorizingAgency documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/authorizationStatement.html}{authorizationStatement documentation}
}
