% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDevelopment.R
\name{ddi_studyDevelopment}
\alias{ddi_studyDevelopment}
\title{studyDevelopment and its child nodes}
\usage{
ddi_studyDevelopment(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Describe the process of study development as a series of development
activities. These activities can be typed using a controlled vocabulary.
More information on these elements, especially their allowed
attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{studyDevelopment} is contained in \code{stdyDscr}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_developmentActivity]{ddi_developmentActivity()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/studyDevelopment.html}{studyDevelopment documentation}
}
