% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_sumStat}
\alias{ddi_sumStat}
\title{DDI Summary Statistics specification}
\usage{
ddi_sumStat(type, ...)
}
\arguments{
\item{type}{The statistics being shown:mean, median, mode, valid cases, invalid cases, minimum, maximum, or standard deviation. If a value of "other" is used here, a value taken from a controlled vocabulary should be put in the "otherType" attribute. This option should only be used when applying a controlled vocabulary to this attribute.}

\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
One or more statistical measures that describe the responses to a particular variable and may include one or more standard summaries, e.g., minimum and
maximum values, median, mode, etc. This element sumStat requires the type attribute with the value of mean, medn, mode, vald, invd, min, max, stdev, or other.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sumStat.html}
}

