% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ddi_table}
\alias{ddi_table}
\title{table and its child nodes}
\usage{
ddi_table(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Used to create a table in DDI 2.5. More information on this element,
especially the allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{table} is contained in the following elements: \code{key}; \code{notes}; \code{otherMat};
and \code{txt}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_tgroup]{ddi_tgroup()}}
\item \code{\link[=ddi_titl]{ddi_titl()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/table.html}{table documentation}
}
