% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_targetSampleSize}
\alias{ddi_targetSampleSize}
\alias{ddi_sampleSize}
\alias{ddi_sampleSizeFormula}
\title{targetSampleSize and its children nodes}
\usage{
ddi_targetSampleSize(...)

ddi_sampleSize(...)

ddi_sampleSizeFormula(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Provides both the target size of the sample (this is the number in the original sample, not the number of respondents) as well as the formula used for determining the sample size.
More information on the allowed attributes and children can be found in the references
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/targetSampleSize.html}{targetSampleSize documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sampleSize.html}{sampleSize documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sampleSizeFormula.html}{sampleSizeFormula documentation}
}
