% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ddi_tbody}
\alias{ddi_tbody}
\title{tbody and its child nodes}
\usage{
ddi_tbody(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
This is the body of the table. More information on this element, especially
the allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{tbody} is contained in \code{tgroup}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_row]{ddi_row()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/tbody.html}{tbody documentation}
}
