% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ddi_tgroup}
\alias{ddi_tgroup}
\alias{ddi_colspec}
\title{tgroup and its child nodes}
\usage{
ddi_tgroup(...)

ddi_colspec(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
This is the table group. More information on this element, especially the
allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{tgroup} is contained in \code{table}.

\emph{tgroup specific child node}
\itemize{
\item \code{ddi_colspec()} is the column specification for each column. It is an
empty element.
}
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_tbody]{ddi_tbody()}}
\item \code{\link[=ddi_thead]{ddi_thead()}}
}
}

\examples{
ddi_tgroup()

# Functions that must be wrapped in ddi_tgroup()

ddi_colspec(align = "left")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/tgroup.html}{tgroup documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/colspec.html}{colspec documentation}
}
