% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ddi_thead}
\alias{ddi_thead}
\title{thead and its child nodes}
\usage{
ddi_thead(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
This is the table header. More information on this element, especially the
allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{thead} is contained in \code{tgroup}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_row]{ddi_row()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/thead.html}{thead documentation}
}
