% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_titl}
\alias{ddi_titl}
\title{titl node}
\usage{
ddi_titl(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
titl is the full authoritative title for the work at the appropriate level:
marked-up document; marked-up document source; study; other material(s)
related to study description; other material(s) related to study. The study
title will in most cases be identical to the title for the marked-up document.
'A full title should indicate the geographic scope of the data collection as
well as the time period covered. Title of data collection
'(codeBook/stdyDscr/citation/titlStmt/titl) maps to Dublin Core Title element.
This element is required in the Study Description citation. More information
on this element, especially its allowed attributes, can be found in the
references.
}
\details{
\emph{Parent nodes}

\code{titl} is contained in the following elements: \code{table} and \code{titlStmt}.
}
\examples{
ddi_titl("Census of Population, 1950 [United States]: Public Use Microdata Sample")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/titl.html}{titl documentation}
}
