% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_titlStmt}
\alias{ddi_titlStmt}
\alias{ddi_titl}
\alias{ddi_subTitl}
\alias{ddi_altTitl}
\alias{ddi_parTitl}
\alias{ddi_IDNo}
\title{titleStmt and its child nodes}
\usage{
ddi_titlStmt(...)

ddi_titl(...)

ddi_subTitl(...)

ddi_altTitl(...)

ddi_parTitl(...)

ddi_IDNo(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Title statement for the work at the appropriate level: marked-up document; marked-up document source; study; study description, other materials; other materials for the study.
More information on the allowed attributes and child nodes can be found below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/titlStmt.html}{titlStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/titl.html}{titl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/subTitl.html}{subTitl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/altTitl.html}{altTitl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/parTitl.html}{parTitl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/IDNo.html}{IDNo documentation}
}
