% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_TotlResp}
\alias{ddi_TotlResp}
\title{DDI Total Responses specifications
The number of responses to this variable. This element might be used if the number of responses does not match added case counts. It may also be used to sum the
frequencies for variable categories.}
\usage{
ddi_TotlResp(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
DDI Total Responses specifications
The number of responses to this variable. This element might be used if the number of responses does not match added case counts. It may also be used to sum the
frequencies for variable categories.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/TotlResp.html}
}

