% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_txt}
\alias{ddi_txt}
\title{DDI Descriptive Text specification}
\usage{
ddi_txt(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Lengthier description of the parent element
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/txt.html}
}

