% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_undocCod}
\alias{ddi_undocCod}
\title{DDI List of Undocumented Codes specification}
\usage{
ddi_undocCod(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Values whose meaning is unknown
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/undocCod.html}
}

