% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_useStmt}
\alias{ddi_useStmt}
\alias{ddi_citReq}
\alias{ddi_conditions}
\alias{ddi_confDec}
\alias{ddi_deposReq}
\alias{ddi_disclaimer}
\alias{ddi_restrctn}
\alias{ddi_specPerm}
\title{useStmt and its children}
\usage{
ddi_useStmt(...)

ddi_citReq(...)

ddi_conditions(...)

ddi_confDec(...)

ddi_deposReq(...)

ddi_disclaimer(...)

ddi_restrctn(...)

ddi_specPerm(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Information on terms of use for the data collection. This element may be repeated only to support multiple language expressions of the content.
For more information on useStmt, its attributes and its children see the references
}
\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_contact]{ddi_contact()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/useStmt.html}{useStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/citReq.html}{citReq documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/conditions.html}{conditions documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/confDec.html}{confDec documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/deposReq.html}{deposReq documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/disclaimer.html}{disclaimer documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/restrctn.html}{restrctn documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/specPerm.html}{specPerm documentation}
}
