% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_var}
\alias{ddi_var}
\title{DDI variable specification}
\usage{
ddi_var(varname, database_id = NULL, ...)
}
\arguments{
\item{varname}{The variable name}

\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
Variables are specified with the var tag, which has a litany of
attributes. The basic necessities of a variable, however, are given by its name, and type.
If the variable is categorical, then the variable will have as many
catgry (\code{ddi_catgry()}) entries as category levels. Missingness codes are specified in \code{ddi_catgry()}.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/var.html}
}

