% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_varFormat}
\alias{ddi_varFormat}
\title{DDI Variable Format specification}
\usage{
ddi_varFormat(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
The technical format of the variable in question. ttributes for this element include: "type," which indicates if the variable is character or numeric;
"formatname," which in some cases may provide the name of the particular, proprietary format actually used; "schema," which identifies the vendor or standards
body that defined the format (acceptable choices are SAS, SPSS, IBM, ANSI, ISO, XML-data or other); "category," which describes what kind of data the format
represents, and includes date, time, currency, or "other" conceptual possibilities; and "URI," which supplies a network identifier for the format definition.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/varFormat.html}
}

