% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_varGrp}
\alias{ddi_varGrp}
\title{DDI varGrp specification}
\usage{
ddi_varGrp(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
A group of variables that may share a common subject, arise from the interpretation of a single question, or are linked by some other factor.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/varGrp.html}
}

