% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_verResp}
\alias{ddi_verResp}
\title{DDI Version Responsibility Statement specification}
\usage{
ddi_verResp(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
The organization or person responsible for the version of the work.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/verResp.html}
}

