% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_version}
\alias{ddi_version}
\title{DDI Version specification
Also known as release or edition. If there have been substantive changes in the data/documentation since their creation, this statement should be
used at the appropriate level.}
\usage{
ddi_version(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
DDI Version specification
Also known as release or edition. If there have been substantive changes in the data/documentation since their creation, this statement should be
used at the appropriate level.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/version.html}
}

