% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_verStmt}
\alias{ddi_verStmt}
\title{DDI Version Statement specification}
\usage{
ddi_verStmt(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
This is the version statement for the work at the appropriate level: marked-up document; marked-up document source; study; study description, other material;
other material for study. A version statement may also be included for a data file, a variable, or an nCube.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/verStmt.html}
}

