% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dots_to_xml_components}
\alias{dots_to_xml_components}
\title{Split variable arguments into content and attributes}
\usage{
dots_to_xml_components(...)
}
\arguments{
\item{...}{Any mix of named and unnamed arguments}
}
\value{
A list with members "attribs" and "content"
}
\description{
Split variable arguments into content and attributes
}
\details{
\subsection{ID attributes}{

Entities in DDI schema can have IDs. As of DDI 3, entities
are requried to have URNs. As this package is an API for
DDI 2.5, it is not a requirement to include entity IDs. If
you'd like to specify them, use the 'id_object' parameter.
See \link{create_object_id} for more details.
}
}
