% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap_content}
\alias{unwrap_content}
\title{Allow text and nodes in ddi_node content}
\usage{
unwrap_content(content)
}
\arguments{
\item{content}{A content list, most likely from \code{dots_to_xml_components()}}
}
\value{
\code{content} contained in an "unwrapping" context
}
\description{
Normally, ddi_node objects can only either have text
content or ddi_node children. This adheres to the XML
schema. However, some elements in DDI have a less strict
content schema, more like HTML, which allow text and XML
tags both as content. This utility is for DDI node authors
who would like to use this feature. Rather than wrap the
content in XML structure, this function marks its input
content as "unwrapped": when passed to \code{as_xml()}, the content
will be directly converted to text, not wrapped in XML.
}
\examples{
if (FALSE) {
  # Building a custom "universe" tag
  # which accepts <txt> and <concept>
  # alongside text
  ddi_universe <- function(...) {
    components <- dots_to_xml_components(...)

    build_branch_node(
      "universe",
      content = unwrap_content(components$content),
      attribs = components$attribs
    )
  }
}
}
