% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_anlyInfo}
\alias{ddi_anlyInfo}
\alias{ddi_dataAppr}
\alias{ddi_EstSmpErr}
\alias{ddi_respRate}
\title{anlyInfo and its child nodes}
\usage{
ddi_anlyInfo(...)

ddi_dataAppr(...)

ddi_EstSmpErr(...)

ddi_respRate(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Information on data appraisal.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/anlyInfo.html}{anylInfo documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataAppr.html}{dataAppr documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/respRate.html}{respRate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/EstSmpErr.html}{EstSmpErr documentation}
}
