% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_anlyInfo}
\alias{ddi_anlyInfo}
\alias{ddi_dataAppr}
\alias{ddi_EstSmpErr}
\alias{ddi_respRate}
\title{anlyInfo and its child nodes}
\usage{
ddi_anlyInfo(...)

ddi_dataAppr(...)

ddi_EstSmpErr(...)

ddi_respRate(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Information on data appraisal.
}
\details{
\emph{Parent nodes}

\code{anlyInfo} is contained in \code{method}.

\emph{anlyInfo specific child nodes}
\itemize{
\item \code{ddi_dataAppr()} are other issues pertaining to data appraisal. Describe
here issues such as response variance, nonresponse rate and testing for bias,
interviewer and response bias, confidence levels, question bias, etc.
Attribute type allows for optional typing of data appraisal processes and
option for controlled vocabulary.
\item \code{ddi_EstSmpErr()} are estimates of sampling error. This element is a
measure of how precisely one can estimate a population value from a given
sample.
\item \code{ddi_respRate()} is the response rate. The percentage of sample members
who provided information. This may include a broader description of
stratified response rates, information affecting resonse rates etc.
}
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/anlyInfo.html}{anylInfo documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataAppr.html}{dataAppr documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/respRate.html}{respRate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/EstSmpErr.html}{EstSmpErr documentation}
}
