% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_anlysUnit}
\alias{ddi_anlysUnit}
\title{anlysUnit node}
\usage{
ddi_anlysUnit(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Provides information regarding whom or what the variable/nCube describes.
The element may be repeated only to support multiple language expressions of
the content. More information on this element, especially its allowed
attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{anlysUnit} is contained in \code{nCube} and \code{var}.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/anlysUnit.html}{anlysUnit documentation}
}
