% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_catgry}
\alias{ddi_catgry}
\title{DDI catgry specification}
\usage{
ddi_catgry(missing = FALSE, ...)
}
\arguments{
\item{missing}{If a category represents a missing response. Default to \code{FALSE}}

\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
A description of a particluar categorical response
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/catgry.html}
}

