% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_citation}
\alias{ddi_citation}
\alias{ddi_sourceCitation}
\alias{ddi_fileCitation}
\alias{ddi_biblCit}
\alias{ddi_holdings}
\title{citation, sourceCitation, fileCitation and their child nodes}
\usage{
ddi_citation(...)

ddi_sourceCitation(...)

ddi_fileCitation(...)

ddi_biblCit(...)

ddi_holdings(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Citation entities for the study including general citations and source citations. More information on citation and
source citation can be found below and in the references.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_distStmt]{ddi_distStmt()}}
\item \code{\link[=ddi_prodStmt]{ddi_prodStmt()}}
\item \code{\link[=ddi_rspStmt]{ddi_rspStmt()}}
\item \code{\link[=ddi_serStmt]{ddi_serStmt()}}
\item \code{\link[=ddi_titlStmt]{ddi_titlStmt()}}
\item \code{\link[=ddi_verStmt]{ddi_verStmt()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/citation.html}{citation documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sourceCitation.html}{sourceCitation documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileCitation.html}{fileCitation documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/serInfo.html}{serInfo documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/holdings.html}{holdings documentation}
}
