% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_citation}
\alias{ddi_citation}
\alias{ddi_titlStmt}
\title{Create a DDI citation}
\usage{
ddi_citation(...)

ddi_titlStmt(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
Citation entities belong in study and DDI document description blocks.
}
