% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_codeBook}
\alias{ddi_codeBook}
\title{Codebook}
\usage{
ddi_codeBook(...)
}
\arguments{
\item{...}{Child nodes or attributes}
}
\description{
The root node of a DDI 2.5 Codebook file. This file must contain stdyDscr
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/codeBook.html}{codeBook documentation}
}
