% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_codingInstructions}
\alias{ddi_codingInstructions}
\alias{ddi_command}
\title{codingInstructions and its child nodes}
\usage{
ddi_codingInstructions(...)

ddi_command(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Describe specific coding instructions used in data processing, cleaning, assession, or tabulation.
More information on codingInstructions allowed attributes and children can be found below and in the references.
}
\section{General children allowed}{

\itemize{
\item \code{ddi_txt()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/codingInstructions.html}{codingInstructions documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/command.html}{command documentation}
}
