% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_concept}
\alias{ddi_concept}
\title{DDI Concept specification}
\usage{
ddi_concept(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
The general subject to which the parent element may be seen as pertaining. This element serves the same purpose as the keywords and topic classification elements,
but at the data description level. The "vocab" attribute is provided to indicate the controlled vocabulary, if any, used in the element, e.g., LCSH (Library of Congress Subject
Headings), MeSH (Medical Subject Headings), etc. The "vocabURI" attribute specifies the location for the full controlled vocabulary.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/concept.html}
}

