% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_contact}
\alias{ddi_contact}
\title{contact node}
\usage{
ddi_contact(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Names and addresses of individuals responsible for the work. Individuals listed as contact persons will be used as resource persons regarding problems or
questions raised by the user community. The URI attribute should be used to indicate a URN or URL for the homepage of the contact individual. The
email attribute is used to indicate an email address for the contact individual.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/contact.html}{contact documentation}
}
