% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccs.R
\name{ddi_dataAccs}
\alias{ddi_dataAccs}
\title{dataAccs and its children}
\usage{
ddi_dataAccs(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
This section describes data access conditions and terms of use for the data collection. In cases where access conditions differ across individual files
or variables, multiple access conditions can be specified.  For more information see the references section.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_setAvail]{ddi_setAvail()}}
\item \code{\link[=ddi_useStmt]{ddi_useStmt()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataAccs.html}{dataAccs documentation}
}
