% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_dataColl}
\alias{ddi_dataColl}
\alias{ddi_actMin}
\alias{ddi_cleanOps}
\alias{ddi_collectorTraining}
\alias{ddi_collMode}
\alias{ddi_collSitu}
\alias{ddi_ConOps}
\alias{ddi_dataCollector}
\alias{ddi_deviat}
\alias{ddi_frequenc}
\alias{ddi_instrumentDevelopment}
\alias{ddi_resInstru}
\alias{ddi_sampProc}
\alias{ddi_timeMeth}
\alias{ddi_weight}
\title{dataColl and its children}
\usage{
ddi_dataColl(...)

ddi_actMin(...)

ddi_cleanOps(...)

ddi_collectorTraining(...)

ddi_collMode(...)

ddi_collSitu(...)

ddi_ConOps(...)

ddi_dataCollector(...)

ddi_deviat(...)

ddi_frequenc(...)

ddi_instrumentDevelopment(...)

ddi_resInstru(...)

ddi_sampProc(...)

ddi_timeMeth(...)

ddi_weight(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Information about the data collection methodology employed in the codebook. More information on dataColl, its attributes, and
its allowed children can be found below and in the references.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_sampleFrame]{ddi_sampleFrame()}}
\item \code{\link[=ddi_sources]{ddi_sources()}}
\item \code{\link[=ddi_targetSampleSize]{ddi_targetSampleSize()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataColl.html}{dataColl documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/actMin.html}{actMin documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/cleanOps.html}{cleanOps documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/collectorTraining.html}{collectorTraining documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/collMode.html}{collMode documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/collSitu.html}{collSitu documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/ConOps.html}{ConOps documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataCollector.html}{dataCollector documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/deviat.html}{deviat documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/frequenc.html}{frequenc documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/instrumentDevelopment.html}{instrumentDevelopment documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/resInstru.html}{resInstru documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sampProc.html}{sampProc documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/timeMeth.html}{timeMeth documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/weight.html}{weight documentation}
}
