% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_dataDscr}
\alias{ddi_dataDscr}
\title{Variable description}
\usage{
ddi_dataDscr(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Description of variables within the Codebook.
}
\section{DDI Codebook 2.5 Documentation}{

\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataDscr.html}
}

