% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_dataFingerprint}
\alias{ddi_dataFingerprint}
\alias{ddi_algorithmSpecification}
\alias{ddi_algorithmVersion}
\alias{ddi_digitalFingerprintValue}
\title{dataFingerprint and its child nodes}
\usage{
ddi_dataFingerprint(...)

ddi_algorithmSpecification(...)

ddi_algorithmVersion(...)

ddi_digitalFingerprintValue(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Allows for assigning a hash value (digital fingerprint) to the data or
data file. More information on these elements, especially their allowed
attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{dataFingerprint} is contained in \code{fileDscr}.

\emph{dataFingerprint specific child nodes}
\itemize{
\item \code{ddi_algorithmSpecification()}
\item \code{ddi_algorithmVersion()}
\item \code{ddi_digitalFingerprintValue()}
}
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataFingerprint.html}{dataFingerprint documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/algorithmSpecification.html}{algorithmSpecification documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/algorithmVersion.html}{algorithmVersion documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/digitalFingerprintValue.html}{digitalFingerprintValue documentation}
}
