% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locMap.R
\name{ddi_dataItem}
\alias{ddi_dataItem}
\alias{ddi_CubeCoord}
\alias{ddi_physLoc}
\title{dataItem and its child nodes}
\usage{
ddi_dataItem(...)

ddi_CubeCoord(...)

ddi_physLoc(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema.}
}
\description{
Identifies a physical storage location for an individual data entry, serving
as a link between the physical location and the logical content description
of each data item. . It is used to describe the physical location of aggregate/tabular
data in cases where the nCube model is employed. More information
on the allowed attributes for dataItem and its child nodes can be found
below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataItem.html}{dataItem documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/CubeCoord.html}{CubeCoord documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/physLoc.html}{physLoc documentation}
}
