% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_dataSrc}
\alias{ddi_dataSrc}
\title{dataSrc node}
\usage{
ddi_dataSrc(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
dataSrc node
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataSrc_1.html}{dataSrc documentation}
}
