% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_embargo}
\alias{ddi_embargo}
\title{embargo node}
\usage{
ddi_embargo(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Provides information on variables/nCubes which are not currently available
because of policies established by the principal investigators and/or data
producers. This element may be repeated to support multiple language
expressions of the content.
}
\details{
\emph{Parent nodes}

\code{embargo} is contained in \code{nCube} and \code{var}.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/embargo.html}{embargo documentation}
}
