% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_embargo}
\alias{ddi_embargo}
\title{DDI Embargo specification}
\usage{
ddi_embargo(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Provides information on variables/nCubes which are not currently available because of policies established by the principal investigators and/or data producers.
The ISO standard for dates (YYYY-MM-DD) is recommended for use with the "date" attribute. An "event" attribute is provided to specify "notBefore" or "notAfter" ("notBefore" is
the default). A "format" attribute is provided to ensure that this information will be machine-processable, and specifies a format for the embargo element.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/embargo.html}
}

