% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_fileDscr}
\alias{ddi_fileDscr}
\title{fileDscr and its children}
\usage{
ddi_fileDscr(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Information about the data file(s) that comprises a collection.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataDscr.html}{dataDscr documentation}
}
