% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_fileStrc}
\alias{ddi_fileStrc}
\title{fileStrc and its child nodes}
\usage{
ddi_fileStrc(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema.}
}
\description{
Type of file structure. More information on the allowed attributes
for fileStrc and its child nodes can be found below and in the references.
}
\details{
#' @section Branch node children allowed:
\itemize{
\item \code{\link[=ddi_recGrp]{ddi_recGrp()}}
}
}
\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileStrc.html}{fileStrc documentation}
}
