% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTxt.R
\name{ddi_fileTxt}
\alias{ddi_fileTxt}
\alias{ddi_dataChck}
\alias{ddi_dataMsng}
\alias{ddi_fileCont}
\alias{ddi_fileName}
\alias{ddi_filePlac}
\alias{ddi_fileType}
\alias{ddi_format}
\alias{ddi_ProcStat}
\title{fileTxt and its leaf node children}
\usage{
ddi_fileTxt(...)

ddi_dataChck(...)

ddi_dataMsng(...)

ddi_fileCont(...)

ddi_fileName(...)

ddi_filePlac(...)

ddi_fileType(...)

ddi_format(...)

ddi_ProcStat(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema. See the references or go the branch node functions for more information.}
}
\description{
Provides descriptive information about the data file. More information
on the allowed attributes for fileTxt and its child nodes can be found
below and in the references.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_dataFingerprint]{ddi_dataFingerprint()}}
\item \code{\link[=ddi_dimensns]{ddi_dimensns()}}
\item \code{\link[=ddi_fileCitation]{ddi_fileCitation()}}
\item \code{\link[=ddi_fileStrc]{ddi_fileStrc()}}
\item \code{\link[=ddi_verStmt]{ddi_verStmt()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{ddi_software()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileTxt.html}{fileTxt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataChck.html}{dataChck documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataMsng.html}{dataMsng documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileCont.html}{fileCont documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileName.html}{fileName documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/filePlac.html}{filePlac documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fileType.html}{fileType documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/format.html}{format documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/ProcStat.html}{ProcStat documentation}
}
