% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_frameUnit}
\alias{ddi_frameUnit}
\alias{ddi_unitType}
\title{frameUnit and its children}
\usage{
ddi_frameUnit(...)

ddi_unitType(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Provides information about the sampling frame unit. More information on frameUnit, its children nodes, and attributes can be found
below and in the references
}
\section{General children allowed}{

\itemize{
\item \code{ddi_txt()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/frameUnit.html}{frameUnit documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/unitType.html}{unitType documentation}
}
