% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_geoBndBox}
\alias{ddi_geoBndBox}
\alias{ddi_eastBL}
\alias{ddi_northBL}
\alias{ddi_southBL}
\alias{ddi_westBL}
\title{geoBndBox and its leaf node children}
\usage{
ddi_geoBndBox(...)

ddi_eastBL(...)

ddi_northBL(...)

ddi_southBL(...)

ddi_westBL(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
The fundamental geometric description for any dataset that models geography. GeoBndBox is the minimum box, defined by west and east
longitudes and north and south latitudes, that includes the largest geographic extent of the dataset's geographic coverage. This
element is used in the first pass of a coordinate-based search. If the boundPoly element is included, then the geoBndBox element MUST be included.
More informaiton on geoBndBox, its attributes and its children can be found in the references
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/geoBndBox.html}{geoBndBox documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/eastBL.html}{eastBL documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/northBL.html}{northBL documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/southBL.html}{southBL documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/westBL.html}{westBL documentation}
}
