% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_labl}
\alias{ddi_labl}
\title{labl node}
\usage{
ddi_labl(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
A short description of the parent element. In the variable label, the length of this phrase may depend on the statistical analysis system
used (e.g., some versions of SAS permit 40-character labels, while some versions of SPSS permit 120 characters), although the DDI itself imposes no restrictions
on the number of characters allowed.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/labl.html}{labl documentation}
}
