% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_location}
\alias{ddi_location}
\title{DDI Location specification}
\usage{
ddi_location(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
This is an empty element containing only attributes. Attributes include "StartPos" (starting position of variable), "EndPos" (ending position of variable), "width" (number
of columns the variable occupies), "RecSegNo" (the record segment number, deck or card number the variable is located on), and "fileid", an IDREF link to the fileDscr element
for the file that this location is within (this is necessary in cases where the same variable may be coded in two different files, e.g., a logical record length type file and
a card image type file). Note that if there is no width or ending position, then the starting position should be the ordinal position in the file, and the file would be
described as free-format. The attribute "locMap" is an IDREF to the element locMap and serves as a pointer to indicate that the location information for the nCube's
cells (aggregate data) is located in that section.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/location.html}
}

