% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_mrow}
\alias{ddi_mrow}
\title{DDI mrow specification}
\usage{
ddi_mrow(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
mrow or Mathematical Row is a wrapper containing the presentation expression mi (ddi_mi()). It creates a single string
without spaces consisting of the individual elements described within it. It can be used to create a single
variable by concatenating other variables into a single string. It is used to create linking variables composed
of multiple non-contiguous parts, or to define unique strings for various category values of a single variable.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/mrow.html}
}

