% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_notes}
\alias{ddi_notes}
\title{DDI Notes and Comments specification}
\usage{
ddi_notes(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
For clarifying information/annotation regarding the parent element
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/notes.html}
}

