% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{ddi_prodStmt}
\alias{ddi_prodStmt}
\alias{ddi_copyright}
\alias{ddi_prodDate}
\alias{ddi_prodPlac}
\alias{ddi_software}
\alias{ddi_fundAg}
\alias{ddi_grantNo}
\title{prodStmt and its child nodes}
\usage{
ddi_prodStmt(...)

ddi_copyright(...)

ddi_prodDate(...)

ddi_prodPlac(...)

ddi_software(...)

ddi_fundAg(...)

ddi_grantNo(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Production statement for the work at the appropriate level: marked-up document; marked-up document source; study; study description, other material;
other material for study. More information on prodStmt, its allowed children and attributes can be found below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/prodStmt.html}{prodStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/copyright.html}{copyright documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/prodDate.html}{prodDate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/prodPlac.html}{prodPlac documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/software.html}{software documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/fundAg.html}{fundAg documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/grantNo.html}{grantNo documentation}
}
