% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_qstn}
\alias{ddi_qstn}
\title{DDI Question specification}
\usage{
ddi_qstn(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
The question element may have mixed content. The element itself may contain text for the question, with the subelements being used to provide further information
about the question. Alternatively, the question element may be empty and only the subelements used. The element has a unique question ID attribute which can be used to link
a variable with other variables where the same question has been asked. This would allow searching for all variables that share the same question ID, perhaps because the questions
was asked several times in a panel design.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/qstn.html}
}

