% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_range}
\alias{ddi_range}
\title{range node}
\usage{
ddi_range(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
This is the actual range of values. The "UNITS" attribute permits
the specification of integer/real numbers. The "min" and "max" attributes
specify the lowest and highest values that are part of the range. The
"minExclusive" and "maxExclusive" attributes specify values that are
immediately outside the range. This is an empty element consisting only of
its attributes. More information on this element, especially its allowed
attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{range} is contained in the following elements: \code{valrng}; \code{invalrng}; and
\code{cohort}.
}
\examples{
ddi_range(min = "1", maxExclusive = "20")

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/range.html}{range documentation}
}
