% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDevelopment.R
\name{ddi_resource}
\alias{ddi_resource}
\title{resource}
\usage{
ddi_resource(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
resource
}
\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_dataSrc]{ddi_dataSrc()}}
\item \code{\link[=ddi_srcOrig]{ddi_srcOrig()}}
\item \code{\link[=ddi_srcDocu]{ddi_srcDocu()}}
\item \code{\link[=ddi_srcOrig]{ddi_srcOrig()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/resource.html}{resource documentation}
}
