% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_respUnit}
\alias{ddi_respUnit}
\title{respUnit node}
\usage{
ddi_respUnit(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Provides information regarding who provided the information contained within
the variable/nCube, e.g., respondent, proxy, interviewer. This element may be
repeated only to support multiple language expressions of the content. More
information on this element, especially its allowed attributes, can be found
in the references.
}
\details{
\emph{Parent nodes}

\code{respUnit} is contained in \code{nCube} and \code{var}.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/respUnit.html}{respUnit documentation}
}
