% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_sampleFrame}
\alias{ddi_sampleFrame}
\alias{ddi_custodian}
\alias{ddi_referencePeriod}
\alias{ddi_sampleFrameName}
\alias{ddi_updateProcedure}
\alias{ddi_validPeriod}
\title{sampleFrame and its children}
\usage{
ddi_sampleFrame(...)

ddi_custodian(...)

ddi_referencePeriod(...)

ddi_sampleFrameName(...)

ddi_updateProcedure(...)

ddi_validPeriod(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Sample frame describes the sampling frame used for identifying the population from which the sample was taken.
More information on sampleFrame, its leaf and branch nodes, and their attributes can be found below and in the
references.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_frameUnit]{ddi_frameUnit()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{ddi_labl()}
\item \code{ddi_txt()}
\item \code{ddi_universe()}
\item \code{ddi_useStmt()}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sampleFrame.html}{sampleFrame documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/custodian.html}{custodian documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/referencePeriod.html}{referencePeriod documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sampleFrameName.html}{sampleFrameName documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/updateProcedure.html}{updateProcedure documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/referencePeriod.html}{referencePeriod documentation}
}
