% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_security}
\alias{ddi_security}
\title{DDI Security specification}
\usage{
ddi_security(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Provides information regarding levels of access, e.g., public, subscriber, need to know. The ISO standard for dates (YYYY-MM-DD) is recommended for use with the
date attribute.
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/security.html}
}

