% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ddi_sources}
\alias{ddi_sources}
\title{sources and its child nodes}
\usage{
ddi_sources(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\value{
A ddi_node object.
}
\description{
Description of sources used for the data collection. The element is nestable
so that the sources statement might encompass a series of discrete source
statements, each of which could contain the facts about an individual source.
This element maps to Dublin Core Source element. More information on this
element, especially its allowed attributes, can be found in the references.
}
\details{
\emph{Parent nodes}

\code{sources} is contained in the following elements: \code{dataColl} and \code{sources}.
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_dataSrc]{ddi_dataSrc()}}
\item \code{\link[=ddi_sourceCitation]{ddi_sourceCitation()}}
\item \code{\link[=ddi_sources]{ddi_sources()}}
\item \code{\link[=ddi_srcChar]{ddi_srcChar()}}
\item \code{\link[=ddi_srcDocu]{ddi_srcDocu()}}
\item \code{\link[=ddi_srcOrig]{ddi_srcOrig()}}
}
}

\examples{
ddi_sources()

}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sources.html}{sources documentation}
}
