% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_standardsCompliance}
\alias{ddi_standardsCompliance}
\alias{ddi_complianceDescription}
\title{standardsCompliance and its leaf node chidlren}
\usage{
ddi_standardsCompliance(...)

ddi_complianceDescription(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
The standards compliance section lists all specific standards complied with during the execution of this study. Specify the standard(s)' name(s) and producer(s) and describe how
the study complied with each standard in complianceDescription. More information can be found below and in the references.
}
\section{Branch node children allowed}{

*\code{\link[=ddi_standard]{ddi_standard()}}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/standardsCompliance.html}{standardsCompliance documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/complianceDescription.html}{complianceDescription documentation}
}
