% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docroot.R
\name{ddi_stdyDscr}
\alias{ddi_stdyDscr}
\title{Study description}
\usage{
ddi_stdyDscr(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
All DDI codebooks must have a study description which
contains information about the study overall, including
a citation to the work(s) referenced in the codebook. At least
one citation must be present, capturing the whole study.
}
