% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_stdyInfo}
\alias{ddi_stdyInfo}
\alias{ddi_abstract}
\alias{ddi_studyBudget}
\title{stdyInfo and its leaf node children}
\usage{
ddi_stdyInfo(...)

ddi_abstract(...)

ddi_studyBudget(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
stdyInfo is the study scope. It contains information about the data collection's scope
across several dimensions, including substantive content, geography, and time. More information
on the allowed attributes for stdyINfo and its child nodes can be found below and in the references.
#'
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_exPostEvaluation]{ddi_exPostEvaluation()}}
\item \code{\link[=ddi_qualityStatmeent]{ddi_qualityStatmeent()}}
\item \code{\link[=ddi_subject]{ddi_subject()}}
\item \code{\link[=ddi_sumDscr]{ddi_sumDscr()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/stdyInfo.html}{stdyInfo documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/abstract.html}{abstract documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/studyBudget.html}{studyBudget documentation}
}
