% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_subject}
\alias{ddi_subject}
\alias{ddi_keyword}
\alias{ddi_topcClas}
\title{subject}
\usage{
ddi_subject(...)

ddi_keyword(...)

ddi_topcClas(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Subject describes the data collection's intellectual content. More information on the allowed attributes and child nodes for subject
can be found below and in the references.
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/subject.html}{subject documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/keyword.html}{keyword documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/topcClas.html}{topcClas documentation}
}
