% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdyInfo.R
\name{ddi_sumDscr}
\alias{ddi_sumDscr}
\alias{ddi_anlyUnit}
\alias{ddi_collDate}
\alias{ddi_dataKind}
\alias{ddi_geogCover}
\alias{ddi_geogUnit}
\alias{ddi_nation}
\alias{ddi_timePrd}
\title{sumDscr and its children}
\usage{
ddi_sumDscr(...)

ddi_anlyUnit(...)

ddi_collDate(...)

ddi_dataKind(...)

ddi_geogCover(...)

ddi_geogUnit(...)

ddi_nation(...)

ddi_timePrd(...)
}
\arguments{
\item{...}{Child nodes or attributes. To set a DDI ID, use \code{id_object} in any \code{ddi_} function to assign the identifier.}
}
\description{
This is the summary data description and it contains information about the geographic coverage of the study and
unit of analysis.
}
\section{Branch node children allowed}{

\itemize{
\item \code{\link[=ddi_boundPoly]{ddi_boundPoly()}}
\item \code{\link[=ddi_geoBndBox]{ddi_geoBndBox()}}
}
}

\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_universe]{ddi_universe()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/sumDscr.html}{sumDscr documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/anlyInfo.html}{anlyInfo documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/collDate.html}{collDate documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/dataKind.html}{dataKind documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/geogCover.html}{geogCover documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/geogUnit.html}{geogUnit documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/nation.html}{nation documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/timePrd.html}{timePrd documentation}
}
