% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_txt}
\alias{ddi_txt}
\title{txt node}
\usage{
ddi_txt(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
Lengthier description of the parent element
}
\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/txt.html}{txt documentation}
}
