% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_useStmt}
\alias{ddi_useStmt}
\alias{ddi_citReq}
\alias{ddi_conditions}
\alias{ddi_confDec}
\alias{ddi_deposReq}
\alias{ddi_disclaimer}
\alias{ddi_restrctn}
\alias{ddi_specPerm}
\title{useStmt and its children}
\usage{
ddi_useStmt(...)

ddi_citReq(...)

ddi_conditions(...)

ddi_confDec(...)

ddi_deposReq(...)

ddi_disclaimer(...)

ddi_restrctn(...)

ddi_specPerm(...)
}
\arguments{
\item{...}{Child nodes or attributes.}
}
\description{
Information on terms of use for the data collection. This element may be
repeated only to support multiple language expressions of the content.
More information on these elements, especially their allowed attributes, can
be found in the references.
}
\details{
\emph{Parent nodes}

\code{useStmt} is contained in the following elements: \code{dataAccs} and \code{sampleFrame}.

\emph{useStmt specific child nodes}
\itemize{
\item \code{ddi_citReq()} is the citation requirement. This is the text of requirement that
a data collection should be cited properly in articles or other publications
that are based on analysis of the data.
\item \code{ddi_conditions()} indicates any additional information that will assist the
user in understanding the access and use conditions of the data collection.
\item \code{ddi_confDec()} is the confidentiality declaration. This element is used to
determine if signing of a confidentiality declaration is needed to access a
resource.
\item \code{ddi_deposReq()} is the deposit requirement. This is information regarding
user responsibility for informing archives of their use of data through
providing citations to the published work or providing copies of the
manuscripts.
\item \code{ddi_disclaimer()} is information regarding responsibility for uses of the
data collection. This element may be repeated to support multiple language
expressions of the content.
\item \code{ddi_restrctn()} are any restrictions on access to or use of the collection
such as privacy certification or distribution restrictions should be
indicated here. These can be restrictions applied by the author, producer,
or disseminator of the data collection. If the data are restricted to only
a certain class of user, specify which type.
\item \code{ddi_specPerm()} is used to determine if any special permissions are required
to access a resource.
}
}
\section{Shared and complex child nodes}{

\itemize{
\item \code{\link[=ddi_contact]{ddi_contact()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/useStmt.html}{useStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/citReq.html}{citReq documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/conditions.html}{conditions documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/confDec.html}{confDec documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/deposReq.html}{deposReq documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/disclaimer.html}{disclaimer documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/restrctn.html}{restrctn documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/specPerm.html}{specPerm documentation}
}
