% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ddi_valrng}
\alias{ddi_valrng}
\title{DDI Range of Valid Data Values Specification}
\usage{
ddi_valrng(...)
}
\arguments{
\item{...}{Any other parameters from the DDI Codebook 2.5 schema}
}
\description{
Values for a particular variable that represent legitimate responses. Must include item or range as a child element
}
\section{DDI Codebook 2.5 Documentation}{


\url{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/valrng.html}
}

