% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ddi_verStmt}
\alias{ddi_verStmt}
\alias{ddi_version}
\alias{ddi_verResp}
\title{verStmt and its child nodes}
\usage{
ddi_verStmt(...)

ddi_version(...)

ddi_verResp(...)
}
\arguments{
\item{...}{Any parameters from the DDI Codebook 2.5 schema}
}
\description{
This is the version statement for the work at the appropriate level: marked-up document; marked-up document source; study; study description, other material;
other material for study. More information on verStmt, its children, and attributes can be found in hte references.
}
\section{General children allowed}{

\itemize{
\item \code{\link[=ddi_notes]{ddi_notes()}}
}
}

\references{
\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/verStmt.html}{verStmt documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/version.html}{version documentation}

\href{https://ddialliance.org/Specification/DDI-Codebook/2.5/XMLSchema/field_level_documentation_files/schemas/codebook_xsd/elements/verResp.html}{verResp documentation}
}
